# [PERF] point_of_sale: cache pricelist rules