@robodoo 2025-03-05
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/200399.png)](https://mergebot.odoo.com/odoo/odoo/pull/200399)