@Jihaysse 2025-03-26
@adgu-odoo Can you r+ back if it's still good for you?
Thanks!